requirements.txt
numpy==2.0.0
pandas==2.2.2
praat-parselmouth==0.4.3
python-dateutil==2.9.0.post0
pytz==2024.1
scipy==1.14.0
six==1.16.0
tzdata==2024.1
